backend/requirements.txt
# Core
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0

# Desktop window
pywebview>=4.4.1

# MAVLink (Phase 5 — install separately if needed)
# pymavlink>=2.4.41

# Build / packaging
pyinstaller>=6.6.0
